optimize async memcached client

+++ libnet/eventloop_group.cc
@@ -5,20 +5,20 @@
 namespace libnet
 {
 using namespace std;
-EventLoopGroup::EventLoopGroup(EventLoop* loop, int num, const std::string& name)
+EventLoopGroup::EventLoopGroup(EventLoop* loop, size_t num, const std::string& name)
   : base_loop_(loop),
     name_(name),
     index_(0),
     num_(num),
     loop_threads_()
 {
-    loop_threads_.reserve(num);
+  loop_threads_.reserve(num);
 };
 
 void EventLoopGroup::start()
 {
   base_loop_->assertInLoopThread();
-  for (int i = 0; i < num_; i++)
+  for (size_t i = 0; i < num_; i++)
   {
     std::string name = name_;
     name.push_back('-');
@@ -34,7 +34,7 @@ EventLoop* EventLoopGroup::getNextLoop()
   base_loop_->assertInLoopThread();
   if (num_ <= 0)
     return base_loop_;
-  int index = index_++ % num_;
+  size_t index = index_++ % num_;
   return loop_threads_[index]->getLoop();
 };
